+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Here is the list of the changes and features provided by the **Rainbow-Node-SDK*
 Warning: Before deploying in production a bot that can generate heavy traffic, please contact ALE.
 All notable changes to Rainbow-Node-SDK will be documented in this file.
 
+## [2.21.0-lts.3] - 2023-09-14
+-   Update file lts_version.
+
 ## [2.21.0-lts.2] - 2023-07-07
 -   Add log in `Conversation` class object.
 -   Fix the update of `Conversation.Messages` when _onReceipt event received.